backend/pyproject.toml
[tool.poetry]
name = "legal-ai-assistant-backend"
version = "0.1.0"
description = "Backend service for the Legal AI Assistant"
authors = ["SimpleLawerBot Team"]
readme = "README.md"
packages = [{ include = "app" }]
include = ["app/prompts/*.txt"]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.110.0"
uvicorn = { extras = ["standard"], version = "^0.29.0" }
python-dotenv = "^1.0.1"
supabase = "^2.3.3"
openai = "^1.3.7"
httpx = "^0.25.2"
numpy = "^1.26.2"
pgvector = "^0.2.3"
python-multipart = "^0.0.9"
tenacity = "^8.2.3"
pydantic-settings = "^2.2.1"
pypdf = "^3.17.4"
python-docx = "^1.1.2"
pytesseract = "^0.3.10"
Pillow = "^10.4.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
pytest-asyncio = "^0.21.1"
httpx = { extras = ["cli"], version = "^0.25.2" }
ruff = "^0.1.7"
mypy = "^1.7.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

